+++ backend/src/server/docs.rs
@@ -1,6 +1,5 @@
 use utoipa::OpenApi;
 
-use super::dtos::{ListVideoDtoPage, MarkerDtoPage, StashVideoDtoPage};
 use super::handlers::library::VideoCleanupResponse;
 use super::handlers::project::{CreateFunscriptBody, ProjectCreateResponse};
 use super::types::*;
