+++ backend/src/server/mod.rs
@@ -1,5 +1,4 @@
 pub mod docs;
-pub mod dtos;
 pub mod error;
 pub mod handlers;
 pub mod static_files;
